+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * change main proeprty of package.json to final build in dist folder, fixes #4
 * use ng-if for search parameter input to avoid rendering issues and performance
 * Hide in use search parameter suggestions, fixed issue #8
+* correctly handle isolation scope of 'placeholder' attribute, fixed #15
 
 ### 1.1.1 - 03 February 2015
 * update README with latest changes to dist files in bower package